+++ core/src/main/java/nl/nn/adapterframework/pipes/FixedForwardPipe.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2013 Nationale-Nederlanden
+   Copyright 2013, 2016 Nationale-Nederlanden
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -15,8 +15,18 @@
 */
 package nl.nn.adapterframework.pipes;
 
+import org.apache.commons.lang.StringUtils;
+
 import nl.nn.adapterframework.configuration.ConfigurationException;
+import nl.nn.adapterframework.core.IPipeLineSession;
+import nl.nn.adapterframework.core.ParameterException;
 import nl.nn.adapterframework.core.PipeForward;
+import nl.nn.adapterframework.core.PipeRunException;
+import nl.nn.adapterframework.core.PipeRunResult;
+import nl.nn.adapterframework.parameters.Parameter;
+import nl.nn.adapterframework.parameters.ParameterList;
+import nl.nn.adapterframework.parameters.ParameterResolutionContext;
+import nl.nn.adapterframework.parameters.ParameterValueList;
 
 /**
  * Provides provides a base-class for a Pipe that has always the same forward.
@@ -30,6 +40,8 @@ import nl.nn.adapterframework.core.PipeForward;
  * <tr><td>{@link #setMaxThreads(int) maxThreads}</td><td>maximum number of threads that may call {@link #doPipe(java.lang.Object, nl.nn.adapterframework.core.IPipeLineSession)} simultaneously</td><td>0 (unlimited)</td></tr>
  * <tr><td>{@link #setDurationThreshold(long) durationThreshold}</td><td>if durationThreshold >=0 and the duration (in milliseconds) of the message processing exceeded the value specified, then the message is logged informatory</td><td>-1</td></tr>
  * <tr><td>{@link #setGetInputFromSessionKey(String) getInputFromSessionKey}</td><td>when set, input is taken from this session key, instead of regular input</td><td>&nbsp;</td></tr>
+ * <tr><td>{@link #setGetInputFromFixedValue(String) getInputFromFixedValue}</td><td>when set, this fixed value is taken as input, instead of regular input</td><td>&nbsp;</td></tr>
+ * <tr><td>{@link #setEmptyInputReplacement(String) emptyInputReplacement}</td><td>when set and the regular input is empty, this fixed value is taken as input</td><td>&nbsp;</td></tr>
  * <tr><td>{@link #setStoreResultInSessionKey(String) storeResultInSessionKey}</td><td>when set, the result is stored under this session key</td><td>&nbsp;</td></tr>
  * <tr><td>{@link #setPreserveInput(boolean) preserveInput}</td><td>when set <code>true</code>, the input of a pipe is restored before processing the next one</td><td>false</td></tr>
  * <tr><td>{@link #setNamespaceAware(boolean) namespaceAware}</td><td>controls namespace-awareness of possible XML parsing in descender-classes</td><td>application default</td></tr>
@@ -50,6 +62,9 @@ import nl.nn.adapterframework.core.PipeForward;
  * 											      <tr><td>T1</td>  <td>error</td></tr>
  *  </table></td><td>&nbsp;</td></tr>
  * <tr><td>{@link #setForwardName(String) forwardName}</td>  <td>name of forward returned upon completion</td><td>"success"</td></tr>
+ * <tr><td>{@link #setSkipOnEmptyInput(String) skipOnEmptyInput}</td><td>when set, this pipe is skipped</td><td>&nbsp;</td></tr>
+ * <tr><td>{@link #setIfParam(String) ifParam}</td><td>when set, this pipe is only executed when the value of parameter with name <code>ifParam</code> equals <code>ifValue</code> (otherwise this pipe is skipped)</td><td>&nbsp;</td></tr>
+ * <tr><td>{@link #setIfValue(String) ifValue}</td><td>see <code>ifParam</code></td><td>&nbsp;</td></tr>
  * </table>
  * </p>
  * <p><b>Exits:</b>
@@ -65,6 +80,9 @@ public class FixedForwardPipe extends AbstractPipe {
 
     private String forwardName = "success";
     private PipeForward forward;
+	private boolean skipOnEmptyInput=false;
+	private String ifParam = null;
+	private String ifValue = null;
     /**
      * checks for correct configuration of forward
      */
@@ -75,7 +93,59 @@ public class FixedForwardPipe extends AbstractPipe {
         if (forward == null)
             throw new ConfigurationException(getLogPrefix(null) + "has no forward with name [" + forwardName + "]");
     }
-	protected PipeForward getForward() {
+
+	public PipeRunResult doInitialPipe(Object input, IPipeLineSession session)
+			throws PipeRunException {
+		if ((input == null || StringUtils.isEmpty(input.toString()))
+				&& isSkipOnEmptyInput()) {
+			return new PipeRunResult(getForward(), input);
+		}
+		if (getIfParam() != null) {
+			boolean skipPipe = true;
+
+			ParameterValueList pvl = null;
+			if (getParameterList() != null) {
+				ParameterResolutionContext prc = new ParameterResolutionContext(
+						(String) input, session);
+				try {
+					pvl = prc.getValues(getParameterList());
+				} catch (ParameterException e) {
+					throw new PipeRunException(this, getLogPrefix(session)
+							+ "exception on extracting parameters", e);
+				}
+			}
+			String ip = getParameterValue(pvl, getIfParam());
+			if (ip == null) {
+				if (getIfValue() == null) {
+					skipPipe = false;
+				}
+			} else {
+				if (getIfValue() != null && getIfValue().equalsIgnoreCase(ip)) {
+					skipPipe = false;
+				}
+			}
+			if (skipPipe) {
+				return new PipeRunResult(getForward(), input);
+			}
+		}
+		return null;
+	}
+
+	private String getParameterValue(ParameterValueList pvl,
+			String parameterName) {
+		ParameterList parameterList = getParameterList();
+		if (pvl != null && parameterList != null) {
+			for (int i = 0; i < parameterList.size(); i++) {
+				Parameter parameter = parameterList.getParameter(i);
+				if (parameter.getName().equalsIgnoreCase(parameterName)) {
+					return pvl.getParameterValue(i).asStringValue(null);
+				}
+			}
+		}
+		return null;
+	}
+
+    protected PipeForward getForward() {
 		return forward;
 	}
  	/**
@@ -87,5 +157,29 @@ public class FixedForwardPipe extends AbstractPipe {
     }
 	public String getForwardName() {
 		return forwardName;
+	}
+
+	public void setSkipOnEmptyInput(boolean b) {
+		skipOnEmptyInput = b;
+	}
+
+	public boolean isSkipOnEmptyInput() {
+		return skipOnEmptyInput;
+	}
+
+	public void setIfParam(String string) {
+		ifParam = string;
+	}
+
+	public String getIfParam() {
+		return ifParam;
+	}
+
+	public void setIfValue(String string) {
+		ifValue = string;
+	}
+
+	public String getIfValue() {
+		return ifValue;
 	}
 }